Moved ControllerWidget to its own file
